requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6
pillow
numpy
opencv-python-headless
tensorflow-cpu; python_version < "3.13"
pydantic[email]>=2.5.3
requests>=2.31.0
sqlite3-binary; platform_system=='Windows' and python_version<"3.10"
sqlalchemy
passlib[argon2]
argon2-cffi
boto3
psycopg2-binary
supabase
python-dotenv
email-validator
